(PC-8132) Move QuitSignupModal to the components folder
